requirements.txt
# Copyright (C) 2023 Intel Corporation
# SPDX-License-Identifier: BSD-3-Clause

absl-py==0.11.0
astor==0.8.1
av==8.0.3
backcall==0.2.0
bleach==1.5.0
cached-property==1.5.2
certifi==2020.6.20
click==7.1.2
colorama==0.4.4
coloredlogs==14.0
cycler==0.10.0
decorator==4.4.2
ffmpeg-python==0.2.0
gast==0.3.3
GPUtil==1.4.0
grpcio==1.33.2
h5py==2.10.0
html5lib==0.9999999
humanfriendly==8.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jsonargparse==2.32.2
jsonnet==0.20.0
jsonschema==3.2.0
Keras==2.1.6
Keras-Applications==1.0.7
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
lmdb==1.1.1
lpips==0.1.4
Markdown==3.3.3
matplotlib==3.3.2
mmcv-full==1.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.2
opencv-python==4.4.0.46
openpyxl==3.0.7
pandas==1.1.4
parso==0.7.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.8.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
scenedetect==0.5.4.1
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
tabulate==0.8.9
tendo==0.2.15
tensorboard==2.2.0
tensorflow==2.2.0
termcolor==1.1.0
tifffile==2020.9.3
torch==1.7.1
torchvision==0.8.2
tqdm==4.51.0
traitlets==4.3.3
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
youtube-dl==2021.1.8
zipp==3.4.0